Fix tests for both python2 and python3
So that was a bit of a bizare adventure. Apparently you cannot use
pytest fixtures with decorators (functools.wraps or unwrapped).

The solution was to pull in a specifically constructed decorator
package and create a decorator with that. This makes the decorators
works, cross python.

Next up we do not have default values for max in py2, so we have to
resort to `iterable or [0]` shenanigans. A bit of an adventure this
one.