Add docker fix for UTF-8 README
So this was a bit of an interesting bug...

Travis failed because I updated the readme (I added rocket emojis if you must know).
This caused the dists to not be installed correctly in docker when running pip. Let
me phrase that again. Docker won't install a python package because I have rocket emojis
in my README. Why? Because it uses ascii to decode the file and by the gods it doesn't
want to fail gracefully. Who would do that? who encodes their source in ascii? Like,
am I being unreasonable here?.